iio: frequency: adf4382: fix en_phase_resync type to propagate errors
regmap_test_bits() returns a negative error code on failure. Using u8
for en_phase_resync truncates any negative return value to a positive
u8, making the subsequent 'if (en_phase_resync < 0)' check always
false and silently ignoring regmap read errors.

Change the type from u8 to int so error codes are properly propagated.

Fixes: 8c537c347c26 ("iio: frequency: adf4382: Add Artemis ADF4382 driver")

+++ drivers/iio/frequency/adf4382.c
@@ -792,7 +792,7 @@ static int adf4382_frac2_compute(struct adf4382_state *st, u64 res,
 				 u32 *mod2_word)
 {
 	u32 channel_spacing;
-	u8 en_phase_resync;
+	int en_phase_resync;
 	u32 chsp_freq;
 	u32 mod2_tmp;
 	u32 mod2_max;
